updated naming convention and prject structure
