+++ src/api/configure.ts
@@ -1,33 +1,15 @@
-import { Config } from '../interfaces';
-
 import { api } from './api';
 import { createItem } from './index';
-import { readNotes } from '../handleJson';
+import { config } from '../disk';
+import { readNotes } from '../disk';
 
-import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'fs';
+import { writeFileSync } from 'fs';
 import * as inquirer from 'inquirer';
 
 const { createPromptModule } = inquirer;
 const prompt = createPromptModule();
 
 import * as os from 'os';
-
-export const config = (): Config => {
-  const config = `${os.homedir()}/jsnotes/config.json`;
-  if (!existsSync(config)) {
-    if (!existsSync(`${os.homedir()}/jsnotes`)) {
-      mkdirSync(`${os.homedir()}/jsnotes`);
-    }
-    writeFileSync(config, JSON.stringify({ token: '', projectId: '' }));
-  }
-  const resolvedConfig = JSON.parse(readFileSync(config).toString());
-  if (!resolvedConfig.token || !resolvedConfig.projectId) {
-    console.log(
-      `You have not completed the configuration for Todoist API connection`
-    );
-  }
-  return resolvedConfig;
-};
 
 export default async () => {
   const { projectId, token } = config();
